Remove unnecessary ID fields (#10)
The numeric solver assigns values to X and Y components of points to place them. IDs are used to label these variables so that the constraint solver can refer to them. So, anything which isn't a variable being solved by the numeric solver doesn't need an ID. This means we can remove the ID fields from geometry entities and lines. Only points (and in the future, distances) will need them.