Benchmark arbitrarily large problem (#27)
Update the criterion benchmark for the "massive parallel line" system. I changed the generator script to take a variable number of parallel lines, so that we can easily generate an arbitrarily large system with code. Then I plugged that into Criterion's [benchmarking with a range of inputs feature](https://bheisler.github.io/criterion.rs/book/user_guide/benchmarking_with_inputs.html#benchmarking-with-a-range-of-values) so we can generate graphs like this:

<img width="994" height="541" alt="Screenshot 2025-08-26 at 4 44 24 PM" src="https://github.com/user-attachments/assets/eadaa14b-a5b9-476d-9e6b-d1fe8637a8f1" />