Massive parallel benchmark was invalid (#29)
As discussed in https://github.com/KittyCAD/ezpz/issues/28,
the residual functions we calculated for the "lines meet at given angle"
constraint (including parallel and perpendicular) assume the lines
do not share a common point. But my parallel benchmark did do that.

So I've replaced it with a new benchmark that generates a lot of vertical
lines. It's still arbitrarily large and can test a big system.