Treat IDs as indices
Now that our variable IDs are just auto-incrementing ints, their
numeric value is always the same as their index/position in the array
of variables. This means we don't need to do a linear search anymore,
ID 2 is always index 2. To put it another way, `index_of` is just the
identity function.

This halves the time spent on each solve, doubling throughput on benchmarks.