Support circle entities and CircleRadius constraints (#62)
This is a (relatively) big PR.

Firstly, in the core `ezpz` library, we add support for constraints on
circles. Only one constraint is added here, constraining a circle's
radius to some fixed length. But we do allow the existing "fix this
point" constraint to work on a circle's center.

Then, in the text representation and the ezpz CLI (which are used for
testing, debugging and benchmarking the core library), we add support
for:

1. declaring circles (e.g. `circle a`, similar to the existing pattern
`point q`)
2. displaying circles in the CLI's text output, or asserting them in
unit tests.
3. support for specifying the circle's radius or center as a constraint.

Here's an example of an ezpz text file showing the new circles, their
constraints, and their outputs:

<img width="869" height="648" alt="Screenshot 2025-09-03 at 8 29 08 PM"
src="https://github.com/user-attachments/assets/53c683f7-ec90-49b2-b7e0-dccf708c9291"
/>