More accurate, simpler benchmarks
The "text representation" shouldn't be benchmarked.
It's just a convenient wrapper so the CLI and test
suite can operate without instanting lots of
complex Rust structs, and non-Rust devs can
use and test ezpz.

So, exclude from benchmark measurement anything
which is only used for the text representation.

Also benchmark the inconsistent case.

Also remove a lot of benchmark boilerplate code,
almost all benchmarks follow the same pattern so
we can abstract it away in a `bench_case` function.