Use consistent vector/point type (#72)
I've made an alias `V` for "2D vector of f64s". Right now it refers to
`nalgebra::Vector2<f64>` but we can easily change it in the future if we
want to use a different crate, or handwritten impl.

I benchmarked this against both docs.rs/cgmath and docs.rs/nalgebra, and
they were both slower. Katie suggests it's because rustc can't inline
across crates.

Regardless, I think having a tiny, handwritten,
mostly-single-lines-of-code implementation will make it easier for rustc
to optimize this code.