Wrap constraints with ID into ConstraintEntry (#150)
Instead of relying on the order constraints come in, give each
constraint an ID, and track it throughout execution.

This is going to be helpful for
https://github.com/KittyCAD/ezpz/pull/147, where the order of each
constraint in the constraint vec given by the user may be reordered,
because the order they're in might not correspond to their prioritized
order.

This should not have any effect on the public API.