Basic freedom analysis (#164)
For now, just tells you if the entire system is underconstrained or not.

If it's underconstrained, you should add more constraints. ezpz will
give you a least-minimum-square error solution.

If it's not underconstrained, it could be overconstrained or
well-constrained. If it's overconstrained, it'll be inconsistent, i.e.
have some unsatisfied constraints. If it's well-constrained, then the
residual should be approximately zero.

In the future, I will specify which input solver variables (from KCL or
other callers) are underconstrained. For now it just checks if the
overall system is underconstrained. TODO!

## How to use

When you want this expensive analysis, just replace
`kcl_ezpz::solve_with_priority` with
`kcl_ezpz::solve_with_priority_analysis`.

Note that this analysis is expensive to run for large systems. Do _NOT_
run it every time the user drags their mouse. Only run it when the user
adds/removes a constraint. Continue using
`kcl_ezpz::solve_with_priority` where user interactivity is important.

## Implementation

There's a trait `Analysis` with two implementations. One does no
analysis, one does degree-of-freedom analysis. The core solver has no
changes, but if a successful solution is found, the final state of the
model is analyzed with whichever `Analysis` trait type was requested. No
analysis is cheap, DOF analysis is expensive. It's up to the user to
choose.

This generic is private and not revealed to users. It only exists to
facilitate code reuse. Instead users will swap between `fn
solve_with_priority` and `fn solve_with_priority_analysis`.