New 'ScalarEqual' constraint (#187)
Used for "these circles have equal radii", or possibly as a replacement
for "these two points are horizontal (same y) or vertical (same x)".

Rather than build a handcrafted test case with the text representation,
I used a property test, to try a lot of different possible formulations.
This works well here where the constraint is very simple to check.