New constraint, PointArcCoincident (#195)
Checks that a point lies on an arc. This is our first constraint with 3
residuals.

1. Does the distance from the arc's center, to the point, equal the
arc's radius?
2. Is the point's angle "above" the arc's start angle?
3. Is the point's angle "below" the arc's end angle?

I am not super confident in the math here, but David is away and I
haven't really got anyone to run it past. I'm putting this up so
Kurt/Jon can start playing with it.